Initial Texture class refactoring. Get rid of API-specific Texture header.
