+++ Source/Urho3D/Graphics/Direct3D9/D3D9Texture2DArray.h
@@ -74,9 +74,11 @@ public:
     /// Return render surface.
     RenderSurface* GetRenderSurface() const { return renderSurface_; }
 
+protected:
+    /// Create the GPU texture.
+    virtual bool Create();
+
 private:
-    /// Create texture array.
-    bool Create();
     /// Handle render surface update event.
     void HandleRenderSurfaceUpdate(StringHash eventType, VariantMap& eventData);
