+++ Source/Urho3D/Graphics/Direct3D9/D3D9Texture3D.h
@@ -65,10 +65,11 @@ public:
     /// Get data from a mip level. The destination buffer must be big enough. Return true if successful.
     bool GetData(unsigned level, void* dest) const;
 
-private:
-    /// Create texture.
-    bool Create();
+protected:
+    /// Create the GPU texture.
+    virtual bool Create();
 
+private:
     /// Image file acquired during BeginLoad.
     SharedPtr<Image> loadImage_;
     /// Parameter file acquired during BeginLoad.
